add the method getCurrentUserInfo inside the onViewCreated inside the TabUserProfileFragment to solve user profile not showing on backpress button
