+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.example.krishanroy.bookswappers"
+        applicationId "com.KrishanRoy.krishanroy.bookswappers"
         minSdkVersion 19
         targetSdkVersion 28
         versionCode 1
