--- app/src/main/java/com/example/krishanroy/bookswappers/ui/GenerateRandomNumber.java
@@ -1,10 +1,0 @@
-package com.example.krishanroy.bookswappers.ui;
-
-import java.util.Random;
-
-public class GenerateRandomNumber {
-    public int randomNumber(){
-        Random random = new Random();
-        return random.nextInt(4000) + 2000;
-    }
-}
