--- app/src/main/java/com/example/krishanroy/bookswappers/ui/HomeScreenFragment.java
@@ -1,218 +1,0 @@
-package com.example.krishanroy.bookswappers.ui;
-
-import android.annotation.SuppressLint;
-import android.content.Context;
-import android.content.Intent;
-import android.net.Uri;
-import android.os.Bundle;
-import android.os.Environment;
-import android.provider.MediaStore;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.Fragment;
-import android.support.v4.content.FileProvider;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SearchView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
-import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.Toast;
-
-import com.example.krishanroy.bookswappers.R;
-import com.example.krishanroy.bookswappers.ui.controller.BookAdapter;
-import com.example.krishanroy.bookswappers.ui.model.Book;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
-import com.jakewharton.rxbinding3.view.RxView;
-
-import java.io.File;
-import java.io.IOException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.LinkedList;
-import java.util.List;
-
-
-public class HomeScreenFragment extends Fragment implements SearchView.OnQueryTextListener {
-    List<Book> bookList;
-    public static final String TAG = "HomeScreenFragment";
-    private BookAdapter bookAdapter;
-    SearchView searchView;
-    private FragmentCommunication listener;
-    private static final int REQUEST_IMAGE_CAPTURE = 1;
-    private ImageView imageView;
-    private String currentPhotoPath;
-    private static final int REQUEST_TAKE_PHOTO = 1;
-    private FloatingActionButton fabCamera;
-
-    private DatabaseReference bookDatabaseReference;
-
-
-    public static HomeScreenFragment newInstance() {
-        return new HomeScreenFragment();
-    }
-
-    public HomeScreenFragment() {
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        if (context instanceof FragmentCommunication) {
-            listener = (FragmentCommunication) context;
-        } else {
-            throw new RuntimeException(context.toString() +
-                    "must implement FragmentCommunication");
-        }
-    }
-
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        ((AppCompatActivity) getActivity()).getSupportActionBar().show();
-    }
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        setHasOptionsMenu(true);
-        return inflater.inflate(R.layout.home_screen_fragment, container, false);
-    }
-
-    @SuppressLint("CheckResult")
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        RecyclerView recyclerView = view.findViewById(R.id.book_recycler_view);
-        searchView = view.findViewById(R.id.home_screen_searchview);
-        searchView.setOnQueryTextListener(this);
-        fabCamera = view.findViewById(R.id.add_books_with_camera_fab);
-        RxView.clicks(fabCamera).subscribe(clicks -> listener.moveToUploadNewBookFragment());
-        bookDatabaseReference = FirebaseDatabase.getInstance().getReference(UploadNewBooksFragment.BOOK_REFERENCE);
-
-        bookList = new LinkedList<>();
-        bookAdapter = new BookAdapter(bookList, listener);
-        recyclerView.setAdapter(bookAdapter);
-        recyclerView.setLayoutManager(new StaggeredGridLayoutManager(2, StaggeredGridLayoutManager.VERTICAL));
-        bookDatabaseReference.addValueEventListener(bookEventListener);
-    }
-
-    ValueEventListener bookEventListener = new ValueEventListener() {
-        @Override
-        public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-            bookList.clear();
-            for (DataSnapshot ds : dataSnapshot.getChildren()) {
-                Book book = ds.getValue(Book.class);
-                bookList.add(book);
-            }
-            bookAdapter.setData(bookList, listener);
-        }
-
-        @Override
-        public void onCancelled(@NonNull DatabaseError databaseError) {
-            Toast.makeText(requireContext(), databaseError.getMessage(), Toast.LENGTH_SHORT).show();
-        }
-    };
-
-
-    private void takePicturesAndSave() {
-        Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-        if (takePictureIntent.resolveActivity(getActivity().getPackageManager()) != null) {
-            File photoFile = null;
-            try {
-                photoFile = createImageFile();
-            } catch (IOException ex) {
-            }
-            if (photoFile != null) {
-                Uri photoURI = FileProvider.getUriForFile(requireContext(),
-                        "com.example.android.fileprovider", photoFile);
-                takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI);
-                startActivityForResult(takePictureIntent, REQUEST_TAKE_PHOTO);
-            }
-        }
-    }
-
-    private File createImageFile() throws IOException {
-        @SuppressLint("SimpleDateFormat") String timeStamp = new SimpleDateFormat("yyyyMMdd_HHmmss").format(new Date());
-        String imageFileName = "JPEG_" + timeStamp + "_";
-        File storageDir = getActivity().getExternalFilesDir(Environment.DIRECTORY_PICTURES);
-        File image = File.createTempFile(
-                imageFileName,
-                ".jpg",
-                storageDir
-        );
-        currentPhotoPath = image.getAbsolutePath();
-        return image;
-    }
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-    }
-
-    @Override
-    public boolean onQueryTextSubmit(String s) {
-        return false;
-    }
-
-    @Override
-    public boolean onQueryTextChange(String s) {
-        List<Book> newBooksList = new LinkedList<>();
-        for (Book b : bookList) {
-            if (b.getUploaderCity().toLowerCase().startsWith(s.toLowerCase()) ||
-                    b.getAuthor().toLowerCase().startsWith(s.toLowerCase()) ||
-                    b.getTitle().toLowerCase().startsWith(s.toLowerCase()) ||
-                    b.getUploaderName().toLowerCase().startsWith(s.toLowerCase())) {
-                newBooksList.add(b);
-            }
-        }
-        bookAdapter.setData(newBooksList, listener);
-        return false;
-    }
-
-    @Override
-    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-
-        inflater.inflate(R.menu.links_menu, menu);
-        super.onCreateOptionsMenu(menu, inflater);
-    }
-
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            /*case R.id.menu_github_link:
-                listener.openTheGitHubLink();
-                break;
-            case R.id.menu_linkedin_link:
-                listener.openTheLinkedInPage();
-                break;*/
-            case R.id.user_profile_menu:
-                listener.moveToUserProfileFragment();
-                //listener.finishFragment(this);
-                break;
-            case R.id.menu_sign_out:
-                listener.signOutFromTheApp();
-                break;
-
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        bookDatabaseReference.removeEventListener(bookEventListener);
-    }
-}
