+++ app/src/main/res/layout/login_screen_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/splash_screen">
+    android:background="@drawable/bookswappers_icon">
 
     <TextView
         android:layout_width="200dp"
