+++ app/src/main/res/xml/file_paths.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <paths xmlns:android="http://schemas.android.com/apk/res/android">
-    <external-path name="my_images" path="Android/data/com.example.krishanroy.bookswappers/files/Pictures" />
+    <external-path name="my_images" path="Android/data/com.KrishanRoy.krishanroy.bookswappers/files/Pictures" />
 </paths>
