--- google-services (1).json
@@ -1,55 +1,0 @@
-{
-  "project_info": {
-    "project_number": "820485487861",
-    "firebase_url": "https://bookswappers-98574.firebaseio.com",
-    "project_id": "bookswappers-98574",
-    "storage_bucket": "bookswappers-98574.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:820485487861:android:631e880dc60b39ea",
-        "android_client_info": {
-          "package_name": "com.example.krishanroy.bookswappers"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "820485487861-h9pli95dlgb5feg1a779uoao3v7kn8ks.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "com.example.krishanroy.bookswappers",
-            "certificate_hash": "61cbd8cd6857fd351548aa06e81f6ead12ef094c"
-          }
-        },
-        {
-          "client_id": "820485487861-qe9ujj0odh0hce49c06qo5oihp183e9g.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyDu2cx22LyYbuJPPX0Z9GH_ZXSi4UO1JWE"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 2,
-          "other_platform_oauth_client": [
-            {
-              "client_id": "820485487861-qe9ujj0odh0hce49c06qo5oihp183e9g.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
